Unable to create duplicate secret names.

+++ presentation/secrets/addsecrethandler.php
@@ -13,14 +13,22 @@ $value = $_POST['Value'];
 $login = $_SESSION['userid'];
 
 $service = new SecretsService();
-$results = $service->addSecrets($login, $secretName, $key, $value);
+$doesExist = $service->doesSecretExist($login, $secretName);
 
-if($results){
-    echo '<div class="container"> Secret was created successfully</div>';
-} else {
-    echo '<div class="container alert alert-danger">There was an error creating the secret.</div>';
+if($doesExist){
+    require_once '../../_header.php';
+    echo '<p>Error: Secret already exists for this user. Please enter a different value.</p>';
+    require_once '../../_footer.php';
+}else {
+
+    $results = $service->addSecrets($login, $secretName, $key, $value);
+    
+    if($results){
+        echo '<div class="container"> Secret was created successfully</div>';
+    } else {
+        echo '<div class="container alert alert-danger">There was an error creating the secret.</div>';
+    }
 }
-
 require_once '../../_footer.php';
 ?>
 </body>
